+++ __init__.py
@@ -2,4 +2,5 @@ from .py import nodes
 
 NODE_CLASS_MAPPINGS = {
     "DiffuseHighSampler": nodes.DiffuseHighSamplerNode,
+    "DiffuseHighParam": nodes.DiffuseHighParamNode,
 }